Modify 'powerline-inactive1' face
'cyan-bg+1' looks too similar to the background (all '*-bg+1' faces are
intended to be used for the large amount of text).  Use 'cyan-bg'
instead: since the inactive mode-line is darker/lighter (for the
light/dark themes accordingly), making the same background as for
'powerline-active1' creates an illusion of different active and inactive
cyan colors.